More pr fixes
