+++ src/features/beregning/beregning-nb.ts
