+++ src/features/beregning/beregning.module.less
@@ -1,0 +1,18 @@
+@import '~/styles/variables.less';
+
+.fradragContainer {
+    margin-bottom: @spacing;
+}
+
+.fradragItemContainer {
+    margin-bottom: @spacing;
+}
+
+.fradragTypeOgBelopContainer {
+    display: flex;
+    justify-content: space-between;
+
+    .fradragtype {
+        margin-right: @spacing-s;
+    }
+}
