+++ src/pages/saksoversikt/vilkår/Beregning.tsx
@@ -5,6 +5,8 @@ import React, { useState } from 'react';
 import DatePicker from 'react-datepicker';
 import { useHistory } from 'react-router-dom';
 
+import messages from '~/features/beregning/beregning-nb';
+import { FradragFormData, isValidFradrag, fradragSchema, FradragInputs } from '~features/beregning';
 import * as sakSlice from '~features/saksoversikt/sak.slice';
 import { toDateOrNull } from '~lib/dateUtils';
 import { useI18n } from '~lib/hooks';
@@ -12,8 +14,6 @@ import { Nullable } from '~lib/types';
 import yup from '~lib/validering';
 import { useAppDispatch } from '~redux/Store';
 
-import messages from '../../saksoversikt/beregning/beregning-nb';
-import { FradragFormData, FradragInputs, fradragSchema, isValidFradrag } from '../beregning/Beregning';
 import VisBeregning from '../beregning/VisBeregning';
 
 import styles from './beregning.module.less';
